fix: just install during install
